Created & exported common & multi-choice functions
